Fix Views custom field handler not displaying select value labels.
The current handler looks up the CiviCRM custom field ID using the field's
label, which is not intended to be unique. More than one field can exist
across multiple different custom field groups with the same field label. If
that happens, when attempting to pull the list of select field options, the
handler will just use the field with the lowest field ID for all fields that
have the same field label, so the field's value won't get rendered.

This change makes the handler use the field's column_name to look up the
field id, which is unique across all fields in CiviCRM.